Add verification process
Add processing of private messages, expecting tokens from a database.
Maches tokens and invalidates them by deleting them after use from the
database.  The expected worflow is that another system, for example a
forum, creates a unique token, with a link to the user it's valid for.
When the token is used, the discord user that gave the token is
authorized by the bot, in addition to the link between the discord user
and the user on the system being stored (currently not implemented).