Add revocation signaling
Add a Unix datagram socket server that listens for signals from the web
applictation.  The protocol is very simple, each datagram is a json
encoded payload with an "action" attribute.  Currently the only action
supported is "revoke", which expects "user_id" to be set to the id of
the Discord user to revoke.

When a revoke request is processed from the web application, a datagram
is sent to the bot over the previously mentioned Unix datagram socket,
to make it take effect imminently.