Handle unregistered users in refresh_id
The refresh_id method is sometimes called for a member that is not
registered.  This can happen for example when an id is revoked, or a
member joins that is not registered.  Add code to properly handle and
revoke the managed roles on mebers that are not regestired when
refresh_id is called on them.