Delay sending user updates
When another DataWriter updates the User class, a nested transaction is
done, which means that the _postSaveAfterTransaction hook is executed
still inside the ongoing transaction.  This means that the update
notification sent to the bot will happen before the data is actually
committed to the database.

Fix by delaying the update notification until after the current
controller has finished dispatching it's action.  This is done by
registring a CodeEvent listener in the _postSave and _postDelete hooks
for the User DataWriter.