Add "Link with Discord" permission
Implement access control for the Discord Sever page and actions via
a new ACL permission, "Link with Discord".  Check for this permission in
the controller for the Discord page, and the controller for handling
actions made on the Discord page.  Add template variable $canLinkDiscord
to the public template (needed for the navigation menu), and to the
wrapper for the account page.  And use this template variable in the
various locations the Discord page is linked, to hide the links if the
current user does not have permission to link with discord.